requirements.txt
openai==1.12.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tqdm==4.65.0 